backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
rasterio
numpy
python-dotenv
requests

